[karst] fix benchmark folder filter
The bnode.sh script searches directories inside `retic_performance`
 for a "valid" benchmark to run.
A "valid" benchmark is a folder with some `karst_input` files in it.

The old filter wasn't working -- every folder was a valid benchmark
folder -- though the next step of finding a worklist was fine.

Using `find` now to check if a folder contains any `karst_input`